Add package graph and move file
